refactor: merge ndk_event_signer_web and web_event_signer_web
