+++ packages/ndk_flutter/lib/signers/src/event_signer_web.dart
@@ -2,11 +2,50 @@ import 'dart:async';
 import 'dart:js_interop';
 
 import 'package:ndk/ndk.dart';
+import 'package:ndk/shared/nips/nip01/bip340.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:web/web.dart' as web;
 
 import '../../verifiers/src/js_interop.dart';
 import 'nostr_sign_js.dart';
+
+/// Web alias of [NdkEventSigner]: fast [WebEventSigner] via JS interop.
+typedef NdkEventSigner = WebEventSigner;
+
+/// Web factory: produces [WebEventSigner] instances using @noble/curves.
+class NdkEventSignerFactory implements LocalEventSignerFactory {
+  const NdkEventSignerFactory();
+
+  @override
+  EventSigner create({String? privateKey, String? publicKey}) {
+    final derivedPublicKey =
+        publicKey ?? (privateKey != null ? derivePublicKey(privateKey) : null);
+
+    if (derivedPublicKey == null) {
+      throw ArgumentError('Either publicKey or privateKey must be provided');
+    }
+
+    return WebEventSigner(
+      privateKey: privateKey,
+      publicKey: derivedPublicKey,
+    );
+  }
+
+  @override
+  String derivePublicKey(String privateKey) => Bip340.getPublicKey(privateKey);
+
+  @override
+  (String, String) generateKeyPair() {
+    final keyPair = Bip340.generatePrivateKey();
+    return (keyPair.privateKey!, keyPair.publicKey);
+  }
+
+  @override
+  EventSigner createWithNewKeyPair() {
+    final (privateKey, publicKey) = generateKeyPair();
+    return create(privateKey: privateKey, publicKey: publicKey);
+  }
+}
 
 /// Web implementation of EventSigner using @noble/curves and Web Crypto API
 /// via JS interop for fast BIP-340 signing, NIP-04 and NIP-44 encryption.
